search.js 파일 분리 / input value 대소문자 구별 가능
