+++ index.html
@@ -34,7 +34,7 @@
       </div> -->
     </div>
     <footer>COPYRIGHT&copy; Hyewon Movies All rights reserved.</footer>
-    <script src="js/api.js"></script>
-    <script src="js/search.js"></script>
+    <script src="js/api.js" type="module"></script>
+    <script src="js/search.js" type="module"></script>
   </body>
 </html>
