+++ js/api.js
@@ -1,3 +1,5 @@
+import { searchMovie } from "./search.js";
+
 const movieCards = document.querySelector(".movie-cards");
 
 const options = {
@@ -42,24 +44,22 @@ function addMovie(movie) {
   movieCards.append(div);
 }
 
-function searchMovie() {
-  const movieCard = movieCards.querySelectorAll(".movie-card");
-  const searchValue = document.getElementById("search__value");
-  const searchBtn = document.querySelector(".search__btn");
+// function searchMovie() {
+//   const movieCard = movieCards.querySelectorAll(".movie-card");
+//   const searchValue = document.getElementById("search__value");
+//   const searchBtn = document.querySelector(".search__btn");
 
-  function handleSearch(e) {
-    e.preventDefault();
-    let value = searchValue.value;
-    movieCard.forEach((element) => {
-      element.classList.remove("hidden");
-      let movieTitle = element.childNodes[2].childNodes[1].innerText;
-      // console.log(movieTitle);
-      // console.log(value);
-      if (!movieTitle.toLowerCase().includes(value)) {
-        element.classList.add("hidden");
-      }
-    });
-  }
-  searchBtn.addEventListener("click", handleSearch);
-}
-test;
+//   function handleSearch(e) {
+//     e.preventDefault();
+//     let value = searchValue.value;
+//     movieCard.forEach((element) => {
+//       element.classList.remove("hidden");
+//       let movieTitle = element.childNodes[2].childNodes[1].innerText;
+
+//       if (!movieTitle.toLowerCase().includes(value.toLowerCase())) {
+//         element.classList.add("hidden");
+//       }
+//     });
+//   }
+//   searchBtn.addEventListener("click", handleSearch);
+// }
