+++ js/search.js
@@ -1,9 +1,21 @@
-// const searchValue = document.getElementById("search__value");
-// const searchBtn = document.querySelector(".search__btn");
+const movieCards = document.querySelector(".movie-cards");
 
-// function handleSearch(e) {
-//   e.preventDefault();
-//   console.log(searchValue.value);
-// }
+export function searchMovie() {
+  const movieCard = movieCards.querySelectorAll(".movie-card");
+  const searchValue = document.getElementById("search__value");
+  const searchBtn = document.querySelector(".search__btn");
 
-// searchBtn.addEventListener("click", handleSearch);
+  function handleSearch(e) {
+    e.preventDefault();
+    let value = searchValue.value;
+    movieCard.forEach((element) => {
+      element.classList.remove("hidden");
+      let movieTitle = element.childNodes[2].childNodes[1].innerText;
+
+      if (!movieTitle.toLowerCase().includes(value.toLowerCase())) {
+        element.classList.add("hidden");
+      }
+    });
+  }
+  searchBtn.addEventListener("click", handleSearch);
+}
